+++ ChangeLog.md
@@ -34,6 +34,7 @@ and simplify code. See the Venice *JMH* benchmark *ReflectionBenchmark.java*
 
 - accidentally bundled OpenSans and SourceCodePro fonts with the Venice 1.7.22 
   Jar 
+- functions `nfirst` and `nlast` to work with Java based lists too